+++ backend/src/cache/cache.module.ts
@@ -3,7 +3,7 @@ import { LogModule } from 'src/log/log.module';
 import { RedisModule } from 'src/redis/redis.module';
 import { CacheRepository } from './repository/cache.repository.interface';
 import { RedisCacheRepository } from './repository/redis-cache.repository';
-import { RedisTTLWorker } from './redis-ttl.worker';
+import { RedisTTLWorker } from '../worker/redis-ttl.worker';
 import { CampaignModule } from 'src/campaign/campaign.module';
 
 @Module({
